php/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\php\requirements.in'
#
click==8.1.7
    # via -r .\php\requirements.in
colorama==0.4.6
    # via click
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r .\php\requirements.in
pillow==10.3.0
    # via -r .\php\requirements.in
pyuca==1.2
    # via -r .\php\requirements.in
smmap==5.0.1
    # via gitdb
